remove AEAD bloat
We don't need to authenticate the encrypted data (with a 16-byte MAC added by Poly1305). It turns
out that UTF-8 is a great way to do that already. The probability of random N bytes being a valid
UTF8 string is 2^-N in the ASCII case (and even less with non-ASCII characters). As a bonus, we get
to test from_bytes() parser much harder by exposting it to random bytes.

There is still the (very rare) scenario where the bytes coincidently form a valid (but wrong) UTF8,
which is not parsable because each line does not have exactly 2 tokens (left assert_eq() and FIXME).