use tabs as delimiter to allow spaces
Using tabs may seem unusual, but it is after all going back to their intended purpose (filling
forms, changing from one field to the next). The advantage is that this frees the space characters,
which can be used in keys and values (and even command names). Sometimes, spaces are even used in
passwords, so allowing the full range of ASCII 32..126 is important here. While there, include
spaces in password generator.

Breaks compatibility, so bump version to 0.2.